Make positions floats... not ints!

+++ db/migrate/20140614151334_create_positions.rb
@@ -1,9 +1,9 @@
 class CreatePositions < ActiveRecord::Migration
   def change
     create_table :positions do |t|
-      t.integer :x, null: false
-      t.integer :y, null: false
-      t.integer :z, null: false
+      t.float :x, null: false
+      t.float :y, null: false
+      t.float :z, null: false
       t.belongs_to :lap, null: false
 
       t.timestamps
